Moved isnumeric to char
