+++ arkouda/numpy/__init__.py
@@ -272,7 +272,8 @@ from arkouda.numpy.pdarraycreation import (
     zeros_like,
 )
 from arkouda.numpy.pdarraymanipulation import append, delete, hstack, vstack
-from arkouda.numpy.strings import Strings, isnumeric
+from arkouda.numpy.strings import Strings
+from arkouda.numpy.char import isnumeric
 from arkouda.numpy.timeclass import (
     Datetime,
     Timedelta,
