--- arkouda/numpy/char/__init__.py
@@ -1,3 +1,0 @@
-from numpy import bool_, character, int_, integer, object_, str_
-
-__all__ = ["bool_", "character", "int_", "integer", "object_", "str_"]
