+++ pytest.ini
@@ -48,6 +48,7 @@ testpaths =
     tests/pandas/join_test.py
     tests/logger_test.py
     tests/message_test.py
+    tests/numpy/char_test.py
     tests/numpy/dtypes_test.py
     tests/numpy/manipulation_functions_test.py
     tests/numpy/array_manipulation_tests.py
